+++ src/Middleware/MiddlewareStack.php
@@ -5,11 +5,12 @@ declare(strict_types=1);
 namespace Maduser\Argon\Routing\Middleware;
 
 use Maduser\Argon\Middleware\Contracts\MiddlewareStackInterface;
+use Psr\Http\Server\MiddlewareInterface;
 
 final readonly class MiddlewareStack implements MiddlewareStackInterface
 {
     /**
-     * @param list<class-string> $middlewares
+     * @param list<class-string<MiddlewareInterface>> $middlewares
      */
     public function __construct(
         private array $middlewares = []
@@ -26,7 +27,7 @@ final readonly class MiddlewareStack implements MiddlewareStackInterface
     }
 
     /**
-     * @return list<class-string>
+     * @return list<class-string<MiddlewareInterface>>
      */
     #[\Override]
     public function toArray(): array